+++ src/library.html
@@ -36,9 +36,7 @@
     <main>
       <section>
         <h1 class="visually-hidden">Library</h1>
-        <div class="container">
           <include src="./partials/movies-container.html"></include>
-        </div>
       </section>
     </main>
     <include src="./partials/footer.html"></include>
